// src/app/Models/Pagos/ValidaValeResponse.ts
export class ValidaValeResponse {
  estatusVale: string;
  nombreDistribuidora: string;
  creditoVale: string;
  estatusDV: string;
  firmaDV: string;
  idDistribuidora: string;
  folioVale: string;
  quincenas: string;
  nombre: string;
  paterno: string;
  materno: string;
  fechaNacimiento: Date;
  ine: string;
  calle: string;
  numero: string;
  colonia: string;
  cP: string;
  municipio: string;
  estado: string;
  estatusCliente: boolean;
  sexo: string;
  DescripcionCodigo: string;
  NumeroCodigo: number;
}