// src/app/Models/Pagos/TablaAmortizacionResponse.ts
export class
TablaAmortizacionResponse {
  DescripcionCodigo: string;
  NumeroCodigo: number;
  capital: number;
  efectivoPuntos: number;
  estatusVale: string;
  fechaPago: string;
  numPago: number;
  puntos: number;
  tipoPago: string;
  total: number;
  totalQuincenal: number;
  quincenas: number;
}